Move visual novel scene code into src/game/scenes/visual_novel
